pyproject.toml
[project]
name = "rag-api-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.134.0",
    "inngest>=0.5.16",
    "llama-index-core>=0.14.15",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.24.0",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.17.0",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
]
